feat: add dataset list page and routing
