+++ package.json
@@ -22,6 +22,7 @@
     "lucide-react": "^0.544.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^7.9.1",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
